Layed out PWM logic
Just need to populate the variables timer1highbits, timer1lowbits,
timer1_20ms_high, and timer1_20ms_low with the correct values for each
case.

Also need to make sure my main loop is properly transitioning after the
sort completes.

Will be well on my way with implementing the main sort algorithm after
this, as the only thing that will remain is interfacing with the IR
sensor and then debugging.