Added EEPROM routines and refactored
Read the datasheet and tried to put together an EEPROM read and EEPROM
write. I then used these to contruct a logging function and a function
to read from the logs. Will now need to integrate these into the rest of
the code and test them.

Learned what the extern and volatile keywords meant and tried to apply
them.