Working EEPROM read/write + tests to prove it
Pretty happy with this. Uncomment the debug flag in main then run sort
on the PIC to see the data and address be printed to the screen. EEPROM
access is working perfectly as desired, rewriting over old logs after
the 4th one. Could extent this to all of the EEPROM in the future but
not now as that would be a nightmare for my UI.

To do:
- Tell the user what log to check for the info of the most recent run
- Test the actual sort algorithm
- Test sensor acquisition functions + actuator functions